+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- :sparkles: Add a text widget.
+
 ## 1.0.1 (2019-03-17)
 - :sparkles: Add a grid container with configurable widgets.
 - :construction_worker_man: Add component file generators.